get list of people functions moved to folder
