+++ routers/__init__.py
@@ -1,0 +1,5 @@
+import imp
+
+
+
+from .getlistofpeople import router as listofpeoplerouter
